Add hermit and docs about hermit (#1659)
Also rewrite some Rush docs from RELEASING to README.

Why hermit: it removes any doubt about which node version devs use, and removes the need for nvm or a global node installation.

Note that the CI builds do not use Hermit. This is because they run on multiple versions of node.